Add files via upload
Code to run serial analysis on high-density EMG data, to classify task labels. It assesses mean accuracy, it lists the set of minimum channels necessary for maximal mean accuracy, it yields the topology of the channel importance in the arrays used for the recordings. It also assesses sensitivity of mean accuracy score to electrode displacement. Data available upon request.